requirements.txt
asgiref==3.7.2
asteval==0.9.31
attrs==23.1.0
azure-core==1.29.3
azure-storage-blob==12.17.0
boxsdk==3.8.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
contourpy==1.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cycler==0.11.0
Cython==3.0.0
dictop==0.2.1
distlib==0.3.6
Django==3.2
django-cors-headers==3.1.0
django-filter==2.4.0
django-storages==1.14
djangorestframework==3.12.4
djangorestframework-filters==1.0.0.dev0
djangorestframework-jwt==1.11.0
drf-api-tracking==1.7.8
drf-extensions==0.7.0
drf-api-tracking==1.7.8
ephem==4.1.4
et-xmlfile==1.1.0
filelock==3.12.2
fonttools==4.42.1
future==0.18.3
gunicorn==20.1.0
h5py==3.9.0
idna==3.4
isodate==0.6.1
itypes==1.2.0
Jinja2==3.1.2
kiwisolver==1.4.5
llvmlite==0.40.1
lmfit==1.0.2
MarkupSafe==2.1.3
matplotlib==3.7.2
NREL-PySAM==4.2.0
numba==0.57.1
numpy==1.24.4
openpyxl==3.0.7
packaging==23.1
pandas==2.0.3
patsy==0.5.3
Pillow==10.0.0
platformdirs==3.5.3
psycopg2-binary==2.9.6
pvlib==0.10.1
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.9
python-barcode==0.13.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
pytz==2023.3.post1
PyYAML==6.0.1
recordclass==0.20
requests==2.31.0
requests-toolbelt==1.0.0
RestrictedPython==6.2
scipy==1.11.2
Shapely==1.8.5.post1
six==1.16.0
solarfactors==1.5.3
sqlparse==0.3.0
statsmodels==0.14.0
typing_extensions==4.7.1
tzdata==2023.3
uncertainties==3.1.7
uritemplate==4.1.1
urllib3==1.26.16
virtualenv==20.23.0
whitenoise==5.2.0